LPS-76536 users-admin-web - moves non-public classes to the internal package
