+++ modules/apps/foundation/users-admin/users-admin-web/src/main/java/com/liferay/users/admin/web/internal/servlet/taglib/ui/navigation/user/UserGeneralScreenNavigationCategory.java
@@ -12,41 +12,39 @@
  * details.
  */
 
-package com.liferay.users.admin.web.servlet.taglib.ui;
+package com.liferay.users.admin.web.internal.servlet.taglib.ui.navigation.user;
 
-import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationEntry;
+import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationCategory;
 import com.liferay.portal.kernel.language.LanguageUtil;
-import com.liferay.portal.kernel.model.User;
-import com.liferay.portal.kernel.util.AggregateResourceBundle;
-import com.liferay.portal.kernel.util.PortalUtil;
-import com.liferay.portal.kernel.util.ResourceBundleUtil;
 import com.liferay.users.admin.web.constants.UserFormConstants;
 
 import java.util.Locale;
-import java.util.ResourceBundle;
+
+import org.osgi.service.component.annotations.Component;
 
 /**
  * @author Pei-Jung Lan
  */
-public abstract class BaseUserScreenNavigationEntry
-	implements ScreenNavigationEntry<User> {
+@Component(
+	property = {"screen.navigation.category.order:Integer=10"},
+	service = ScreenNavigationCategory.class
+)
+public class UserGeneralScreenNavigationCategory
+	implements ScreenNavigationCategory {
+
+	@Override
+	public String getCategoryKey() {
+		return UserFormConstants.CATEGORY_KEY_GENERAL;
+	}
 
 	@Override
 	public String getLabel(Locale locale) {
-		return LanguageUtil.get(getResourceBundle(locale), getEntryKey());
+		return LanguageUtil.get(locale, "general");
 	}
 
 	@Override
 	public String getScreenNavigationKey() {
 		return UserFormConstants.SCREEN_NAVIGATION_KEY_USERS;
-	}
-
-	protected ResourceBundle getResourceBundle(Locale locale) {
-		ResourceBundle bundleResourceBundle = ResourceBundleUtil.getBundle(
-			"content.Language", locale, getClass());
-
-		return new AggregateResourceBundle(
-			bundleResourceBundle, PortalUtil.getResourceBundle(locale));
 	}
 
 }
